(PC-17806) refactor(auth): move all suspension screen in pages folder
