+++ src/features/auth/api/useAccountUnsuspend.test.ts
@@ -1,10 +1,11 @@
 import { rest } from 'msw'
 
-import { useAccountUnsuspend } from 'features/auth/suspendedAccount/SuspendedAccount/useAccountUnsuspend'
 import { env } from 'libs/environment'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { server } from 'tests/server'
 import { renderHook, waitFor } from 'tests/utils'
+
+import { useAccountUnsuspend } from './useAccountUnsuspend'
 
 const onSuccess = jest.fn()
 const onError = jest.fn()
