+++ src/features/auth/pages/suspendedAccount/SuspendedAccount/SuspendedAccount.tsx
@@ -2,10 +2,10 @@ import { useNavigation } from '@react-navigation/native'
 import React from 'react'
 import styled from 'styled-components/native'
 
+import { useAccountSuspensionDate } from 'features/auth/api/useAccountSuspensionDate'
+import { useAccountUnsuspend } from 'features/auth/api/useAccountUnsuspend'
 import { useSettingsContext } from 'features/auth/context/SettingsContext'
 import { useLogoutRoutine } from 'features/auth/logout/useLogoutRoutine'
-import { useAccountSuspensionDate } from 'features/auth/suspendedAccount/SuspendedAccount/useAccountSuspensionDate'
-import { useAccountUnsuspend } from 'features/auth/suspendedAccount/SuspendedAccount/useAccountUnsuspend'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { analytics } from 'libs/firebase/analytics'
